% Generated by roxygen2 (4.0.0): do not edit by hand
\name{antipode}
\alias{antipode}
\alias{antipode.Coordinate}
\alias{antipode.default}
\title{Calculate antipodes}
\usage{
antipode(...)

\method{antipode}{default}(latitude, longitude, ...)

\method{antipode}{Coordinate}(coordinate, ...)
}
\arguments{
\item{latitude}{a latitude coordinate}

\item{longitude}{a longitude coordinate}

\item{coordinate}{a coordinate class}

\item{...}{other arguments}
}
\description{
Calculate the antipodes of the idicated point.
}

