% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Coordinate}
\alias{Coordinate}
\title{Create a Coordinate object}
\usage{
Coordinate(latitude, longitude)
}
\arguments{
\item{latitude}{a latitude coordinate}

\item{longitude}{a longitude coordinate}
}
\description{
The method create a new Coordinate object type using the indicated
latitude and longitude
}

