\name{Coordinates}
\alias{Coordinates}
\title{Create a Coordinates object}
\usage{
  Coordinates(latitude, longitude)
}
\arguments{
  \item{latitude}{a latitude coordinate}

  \item{longitude}{a longitude coordinate}
}
\description{
  The method create a new Coordinates object type using the
  indicated latitude and longitude
}

