% Generated by roxygen2 (4.0.0): do not edit by hand
\name{deg2any}
\alias{deg2any}
\title{Convert to a radians or degrees}
\usage{
deg2any(coordinate, units = "degrees")
}
\arguments{
\item{coordinate}{degrees to convert}

\item{units}{a string with the units (degrees or radians)}
}
\description{
Convert from degrees to radians or degrees
}

