% Generated by roxygen2 (4.0.0): do not edit by hand
\name{dms2deg}
\alias{dms2deg}
\title{Convert degrees, minutas and seconds to degrees}
\usage{
dms2deg(degrees, minutes, seconds)
}
\arguments{
\item{degrees}{degrees}

\item{minutes}{minutes}

\item{seconds}{seconds}
}
\value{
the position using degrees in decimal format
}
\description{
Transform the values in degrees, minutes and seconds to degrees in decimal
format
}

