\name{dms2rad}
\alias{dms2rad}
\title{Convert degrees, minutas and seconds to radians}
\usage{
  dms2rad(degrees, minutes, seconds)
}
\arguments{
  \item{degrees}{degrees}

  \item{minutes}{minutes}

  \item{seconds}{seconds}
}
\value{
  the position using degrees in radians
}
\description{
  Transform the values in degrees, minutes and seconds to
  radians
}

