\name{earthDiameter}
\alias{earthDiameter}
\title{Returns the diameter of the earth}
\usage{
  earthDiameter(units = "km")
}
\arguments{
  \item{units}{a string with the distance units (default
  kilometers)}
}
\description{
  Returns the value used for the diameter of the earth in
  the calculations
}

