% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.connected}
\alias{is.connected}
\title{Check is the computer has access to google}
\usage{
is.connected()
}
\description{
Return TRUE where the computer can access to Google
}
\examples{
if (is.connected()) {
  message('The computer can access to Google')
}
}

