% Generated by roxygen2 (4.0.0): do not edit by hand
\name{is.windows}
\alias{is.windows}
\title{Check is the computer is runing Windows}
\usage{
is.windows()
}
\description{
Returns TRUE for computers runing Windows OS
}
\examples{
if (is.windows()) {
  message('The computer is running Windows OS')
}
}

