% Generated by roxygen2 (4.0.0): do not edit by hand
\name{validate.Latitude}
\alias{validate.Latitude}
\title{Validate latitude values}
\usage{
validate.Latitude(latitude)
}
\arguments{
\item{latitude}{a latitude coordinate}
}
\description{
Set the input latitude values between -90 and +90 degrees
}

