% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{whereiam-package}
\alias{whereiam}
\alias{whereiam-package}
\title{whereiam: Geolocation package and spherical trigonometry calculations}
\description{
Provides an interface to different geolocation web API and functions
for spherical trigonometry calculations on Earth.
}
\details{
\tabular{ll}{
  Package: \tab whereiam \cr
  Version: \tab 0.0.6 \cr
  License: \tab GPL-3 \cr
}
}
\author{
Daniel Rodriguez Perez \email{daniel.rodriguez.perez@gmail.com}
}
\references{
http://www.movable-type.co.uk/scripts/latlong.html

http://www.movable-type.co.uk/scripts/latlong-vincenty.html

http://en.wikipedia.org/wiki/Great-circle_distance
}

