\docType{package}
\name{whereiam}
\alias{whereiam}
\alias{whereiam-package}
\title{whereiam: Geolocation package for R}
\description{
  Provides an interface to the geolocation web API
}
\details{
  \tabular{ll}{ Package: \tab whereiam \cr Version: \tab
  0.0.1 \cr License: \tab GPL-3 \cr }
}
\author{
  Daniel Rodriguez Perez
  \email{daniel.rodriguez.perez@gmail.com}
}

