% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getExternalIP}
\alias{getExternalIP}
\title{Get the external IP of the computer}
\usage{
getExternalIP(service = "whatismyipaddress")
}
\arguments{
\item{service}{the optional service used to obtain the external IP}
}
\description{
Returns the external IP of the computer if it is connected to internet,
otherwise returns NULL

The list of services available are: whatismyipaddress, ifconfig, icanhazip,
ipecho and appspot
}

