% Generated by roxygen2 (4.0.0): do not edit by hand
\name{intersection}
\alias{intersection}
\alias{intersection.GreatCircle}
\title{Intersection of two great circle}
\usage{
intersection(...)

\method{intersection}{GreatCircle}(greatCircle1, greatCircle2, ...)
}
\arguments{
\item{greatCircle1}{a GreatCircle class}

\item{greatCircle2}{a GreatCircle class}

\item{...}{other argument}
}
\description{
Calculate of intersection of two great circle paths
}
\references{
http://www.movable-type.co.uk/scripts/latlong.html
}

